add opponent module
